+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The most recent changes are on top, in each type of changes category.
 
 
 ### Fixed
+- Fix no search results info [#601](https://github.com/cyfronet-fid/sat4envi/issues/601)
 
 
 ### Updated